Events: Organize event query functions into a dedicated file
